+++ src/tor_ingress.rs
@@ -31,8 +31,9 @@ use serde::{Deserialize, Serialize};
 use crate::{
     Error, Result,
     kubernetes::{
-        self, Annotations, Api, ConditionsExt, Labels, Object, Resource as KubernetesResource,
-        ResourceName, Torrc as KubernetesTorrc, error_policy, pod_security_context,
+        self, Annotations, Api, ConditionsExt, ErrorBackoff, Labels, Object,
+        Resource as KubernetesResource, ResourceName, Torrc as KubernetesTorrc, error_policy,
+        pod_security_context,
     },
     metrics::Metrics,
     onion_balance::{
@@ -1018,6 +1019,7 @@ pub async fn run_controller(client: Client, config: Config, metrics: Metrics) {
         Arc::new(Context {
             client,
             _config: config,
+            error_backoff: ErrorBackoff::default(),
             metrics,
         }),
     )
@@ -1033,10 +1035,15 @@ pub async fn run_controller(client: Client, config: Config, metrics: Metrics) {
 struct Context {
     client: Client,
     _config: Config,
+    error_backoff: ErrorBackoff,
     metrics: Metrics,
 }
 
 impl kubernetes::Context for Context {
+    fn error_backoff(&self) -> &ErrorBackoff {
+        &self.error_backoff
+    }
+
     fn metrics(&self) -> &Metrics {
         &self.metrics
     }
@@ -1184,6 +1191,8 @@ async fn reconciler(object: Arc<TorIngress>, ctx: Arc<Context>) -> Result<Action
         &state,
     )
     .await?;
+
+    ctx.error_backoff.reset(object.as_ref());
 
     tracing::info!("reconciled");